+++ src/test/java/com/mitchellbosecke/pebble/CoreTagsTest.java
@@ -23,6 +23,7 @@ import java.util.*;
 import java.util.concurrent.Executors;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.fail;
 
 public class CoreTagsTest extends AbstractTest {
 
@@ -139,6 +140,21 @@ public class CoreTagsTest extends AbstractTest {
 
 
     @Test
+    public void testIfWhenInvalidOrNoEndifTag() throws PebbleException, IOException {
+        PebbleEngine pebble = new PebbleEngine.Builder().loader(new StringLoader()).strictVariables(false).build();
+        String source = "{% if variable %}smth{ endif %}";
+        try {
+            pebble.getTemplate(source);
+
+            fail("Should fail due to invalid endif tag");
+        } catch (PebbleException ex) {
+            assertEquals(ex.getPebbleMessage(), "Unexpected end of template. Pebble was looking for the \"endif\" tag");
+            assertEquals(ex.getLineNumber(), (Integer) 1);
+            assertEquals(ex.getFileName(), source);
+        }
+    }
+
+    @Test
     public void testFlush() throws PebbleException, IOException {
         PebbleEngine pebble = new PebbleEngine.Builder().loader(new StringLoader()).strictVariables(false).build();
 
@@ -213,6 +229,21 @@ public class CoreTagsTest extends AbstractTest {
 
         Writer writer = new StringWriter();
         template.evaluate(writer, context);
+    }
+
+    @Test
+    public void testForWhenInvalidOrNoEndforTag() throws PebbleException, IOException {
+        PebbleEngine pebble = new PebbleEngine.Builder().loader(new StringLoader()).strictVariables(false).build();
+        String source = "{% for i in 'a'..5 %}{{i}}% endfor %}";
+        try {
+            pebble.getTemplate(source);
+
+            fail("Should fail due to invalid endfor tag");
+        } catch (PebbleException ex) {
+            assertEquals(ex.getPebbleMessage(), "Unexpected end of template. Pebble was looking for the \"endfor\" tag");
+            assertEquals(ex.getLineNumber(), (Integer) 1);
+            assertEquals(ex.getFileName(), source);
+        }
     }
 
     @Test